react-app: refactor proposal history to be more reusable refs oursky/likedao#234
