--- react-app/src/components/PortfolioScreen/PortfolioScreen.graphql
@@ -1,30 +1,0 @@
-query PortfolioScreenQuery(
-  $first: Int!
-  $after: Int!
-  $isVoter: Boolean!
-  $isSubmitter: Boolean!
-  $isDepositor: Boolean!
-  $address: String!
-  $searchTerm: String
-  $order: ProposalSort
-) {
-  proposalVotesDistribution(address: $address) {
-    ...PortfolioHistoryProposalVotesDistribution
-  }
-  proposals(
-    input: {
-      first: $first
-      after: $after
-      address: {
-        address: $address
-        isVoter: $isVoter
-        isSubmitter: $isSubmitter
-        isDepositor: $isDepositor
-      }
-      searchTerm: $searchTerm
-      order: $order
-    }
-  ) {
-    ...PortfolioHistoryProposalConnection
-  }
-}
